llm_context_manager/requirements.txt
openai==1.3.0
python-dotenv==1.0.0
tiktoken==0.5.1
google-generativeai==0.3.1
flask==3.0.0
werkzeug==3.0.1